Update and rename main.py to game.py
